refactor(acl): rename some of the methods
