+++ src/framework/security/index.ts
@@ -6,5 +6,5 @@
 export * from './security.options';
 export * from './security.module';
 export * from './services/acl.service';
-export * from './services/authorization-checker.service';
+export * from './services/access-checker.service';
 export * from './services/role.provider';
